backend/requirements.txt
Django==4.2.13
djangorestframework==3.15.1
djoser==2.2.3
django-filter==2.4.0
pillow==10.3.0
python-dotenv==1.0.1
drf-spectacular==0.27.2
gunicorn==22.0.0
django-cleanup==8.1.0
drf_extra_fields==3.7.0
django-phonenumber-field[phonenumbers]==7.3.0
numpy==1.26.4
catboost==1.2.5
celery==5.4.0
redis==5.1.0b7
flower==2.0.1
psycopg2-binary==2.9.9
django-cors-headers==4.4.0